aggiunta versione windows del sw esempio proxy

+++ StructuralPatterns/src/main/java/proxy/ProxyPatternTest.java
@@ -2,15 +2,47 @@ package proxy;
 
 public class ProxyPatternTest {
 
+	private static String OS = System.getProperty("os.name").toLowerCase();
+
 	public static void main(String[] args) {
 
 		CommandExecutor executor = new CommandExecutorProxy("scelia", "password_sbagliata");
 
 		try {
-			executor.runCommand("ls -ltr");
-			executor.runCommand(" rm -rf --no-preserve-root /");
+			if (isWindows()) {
+				executor.runCommand("cmd.exe /c cd .");
+				executor.runCommand("cmd.exe /c format c:\\");
+			} else {
+				executor.runCommand("ls -ltr");
+				executor.runCommand(" rm -rf --no-preserve-root /");
+			}
+
 		} catch (Exception exception) {
 			System.out.println(exception.getMessage());
 		}
+	}
+
+	public static boolean isWindows() {
+
+		return (OS.indexOf("win") >= 0);
+
+	}
+
+	public static boolean isMac() {
+
+		return (OS.indexOf("mac") >= 0);
+
+	}
+
+	public static boolean isUnix() {
+
+		return (OS.indexOf("nix") >= 0 || OS.indexOf("nux") >= 0 || OS.indexOf("aix") > 0 );
+
+	}
+
+	public static boolean isSolaris() {
+
+		return (OS.indexOf("sunos") >= 0);
+
 	}
 }
